flask-ml/requirements.txt
Flask==2.3.3
scikit-learn==1.7.0
joblib==1.4.2
spacy==3.7.4
python-dotenv==1.0.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
PyJWT